+++ lib/accept_headers/media_type/negotiator.rb
@@ -6,7 +6,7 @@ module AcceptHeaders
     class Negotiator
       include Negotiatable
 
-      PARAM_PATTERN = /(?<attribute>[\w!#$%^&*\-\+{}\\|'.`~]+)\s*\=\s*(?:\"(?<value>[^"]*)\"|\'(?<value>[^']*)\'|(?<value>[\w!#$%^&*\-\+{}\\|\'.`~]*))/
+      PARAMS_PATTERN = /(?<attribute>[\w!#$%^&*\-\+{}\\|'.`~]+)\s*\=\s*(?:\"(?<value>[^"]*)\"|\'(?<value>[^']*)\'|(?<value>[\w!#$%^&*\-\+{}\\|\'.`~]*))/
 
       private
       def parse(original_header)
@@ -16,7 +16,7 @@ module AcceptHeaders
         return [MediaType.new] if header.empty?
         media_types = []
         header.split(',').each do |entry|
-          accept_media_range, accept_params = entry.split(';', 2)
+          accept_media_range, accept_extensions = entry.split(';', 2)
           next if accept_media_range.nil?
           media_range = MediaType::MEDIA_TYPE_PATTERN.match(accept_media_range)
           next if media_range.nil?
@@ -24,8 +24,8 @@ module AcceptHeaders
             media_types << MediaType.new(
               media_range[:type],
               media_range[:subtype],
-              q: parse_q(accept_params),
-              params: parse_params(accept_params)
+              q: parse_q(accept_extensions),
+              extensions: parse_extensions(accept_extensions)
             )
           rescue Error
             next
@@ -34,19 +34,19 @@ module AcceptHeaders
         media_types.sort! { |x,y| y <=> x }
       end
 
-      def parse_params(params_string)
-        return {} if !params_string || params_string.empty?
-        if params_string.match(/['"]/)
-          params = params_string.scan(PARAM_PATTERN).map(&:compact).to_h
+      def parse_extensions(extensions_string)
+        return {} if !extensions_string || extensions_string.empty?
+        if extensions_string.match(/['"]/)
+          extensions = extensions_string.scan(PARAMS_PATTERN).map(&:compact).to_h
         else
-          params = {}
-          params_string.split(';').each do |part|
-            param = PARAM_PATTERN.match(part)
-            params[param[:attribute]] = param[:value] if param
+          extensions = {}
+          extensions_string.split(';').each do |part|
+            param = PARAMS_PATTERN.match(part)
+            extensions[param[:attribute]] = param[:value] if param
           end
         end
-        params.delete('q')
-        params
+        extensions.delete('q')
+        extensions
       end
     end
   end